Move snackbar to app.js to make it global
